support phase 1 fp name
