+++ src/app/components/Staking/FinalityProviders/FinalityProviders.tsx
@@ -9,18 +9,11 @@ import { FinalityProviderFilter } from "./FinalityProviderFilter";
 import { FinalityProviderSearch } from "./FinalityProviderSearch";
 import { FinalityProviderTable } from "./FinalityProviderTable";
 
-export const FinalityProviders = () => {
+const FinalityProviderContent = () => {
   const { setValue } = useFormContext();
 
   return (
-    <div className="flex flex-col gap-4">
-      <Heading variant="h5" className="text-primary-dark">
-        Step 1
-      </Heading>
-      <Text variant="body1" className="text-primary-dark">
-        Select a Finality Provider
-      </Text>
-
+    <>
       <div className="flex flex-col md:flex-row gap-4">
         <div className="flex-1">
           <FinalityProviderSearch />
@@ -41,6 +34,21 @@ export const FinalityProviders = () => {
       />
 
       <HiddenField name="finalityProvider" defaultValue="" />
+    </>
+  );
+};
+
+export const FinalityProviders = () => {
+  return (
+    <div className="flex flex-col gap-4">
+      <Heading variant="h5" className="text-primary-dark">
+        Step 1
+      </Heading>
+      <Text variant="body1" className="text-primary-dark">
+        Select a Finality Provider
+      </Text>
+
+      <FinalityProviderContent />
     </div>
   );
 };
