+++ src/app/hooks/client/api/useFinalityProvidersV2.ts
@@ -3,13 +3,13 @@ import { useEffect } from "react";
 
 import {
   type PaginatedFinalityProviders,
-  getFinalityProviders,
-} from "@/app/api/getFinalityProviders";
+  getFinalityProvidersV2,
+} from "@/app/api/getFinalityProvidersV2";
 import { ONE_MINUTE } from "@/app/constants";
 import { useError } from "@/app/context/Error/ErrorContext";
 import { ErrorState } from "@/app/types/errors";
 
-const FINALITY_PROVIDERS_KEY = "GET_FINALITY_PROVIDERS_V1_KEY";
+const FINALITY_PROVIDERS_KEY = "GET_FINALITY_PROVIDERS_V2_KEY";
 
 interface Params {
   pk?: string;
@@ -18,13 +18,18 @@ interface Params {
   order?: "asc" | "desc";
 }
 
-export function useFinalityProviders({ pk, sortBy, order, name }: Params = {}) {
+export function useFinalityProvidersV2({
+  pk,
+  sortBy,
+  order,
+  name,
+}: Params = {}) {
   const { isErrorOpen, handleError, captureError } = useError();
 
   const query = useInfiniteQuery({
     queryKey: [FINALITY_PROVIDERS_KEY],
     queryFn: ({ pageParam = "" }) =>
-      getFinalityProviders({ key: pageParam, pk, sortBy, order, name }),
+      getFinalityProvidersV2({ key: pageParam, pk, sortBy, order, name }),
     getNextPageParam: (lastPage) =>
       lastPage?.pagination?.next_key !== ""
         ? lastPage?.pagination?.next_key
