+++ src/app/types/finalityProviders.ts
@@ -1,6 +1,6 @@
 export interface FinalityProvider {
   description: Description;
-  state: FinalityProviderState;
+  state: "active" | "standby";
   commission: string;
   btcPk: string;
   activeTVLSat: number;
@@ -16,20 +16,3 @@ export interface Description {
   securityContact: string;
   details: string;
 }
-
-export enum FinalityProviderState {
-  ACTIVE = "FINALITY_PROVIDER_STATUS_ACTIVE",
-  INACTIVE = "FINALITY_PROVIDER_STATUS_INACTIVE",
-  JAILED = "FINALITY_PROVIDER_STATUS_JAILED",
-  SLASHED = "FINALITY_PROVIDER_STATUS_SLASHED",
-}
-
-export const FinalityProviderStateLabels: Record<
-  FinalityProviderState,
-  string
-> = {
-  [FinalityProviderState.ACTIVE]: "Active",
-  [FinalityProviderState.INACTIVE]: "Inactive",
-  [FinalityProviderState.JAILED]: "Jailed",
-  [FinalityProviderState.SLASHED]: "Slashed",
-};
